% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Compare predictions and labels}
\usage{
compare(
  predictions,
  labels,
  partition = TRUE,
  names = c(specificity = 0, sensitivity = 1)
)
}
\arguments{
\item{predictions}{}

\item{labels}{}

\item{partition}{}

\item{names}{}
}
\value{

}
\description{
Compare predictions and labels
}
