% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R, R/define_contrast.R
\name{define_contrast}
\alias{define_contrast}
\title{Define a contrast (matrix) to specify exact hypothesis system}
\usage{
define_contrast(type = c("raw", "Dunnett", "Tukey"), comparator = NA)

define_contrast(type = c("raw", "Dunnett", "Tukey"), comparator = NA)
}
\arguments{
\item{type}{character, either "Raw", "Dunnett" or "Tukey")}

\item{comparator}{either integer (index of comparator) or character (name of comparator)}
}
\value{
\code{DTAmc_contrast} object to be passed to \code{\link{dta}}

\code{DTAmc_contrast} object to be passed to \code{\link{dta}}
}
\description{
Define a contrast (matrix) to specify exact hypothesis system

Define a contrast (matrix) to specify exact hypothesis system
}
\examples{
define_contrast("Dunnett", 1)
define_contrast("Dunnett", 1)
}
