% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dta.R
\name{dta}
\alias{dta}
\title{Diagnostic test accuracy analysis}
\usage{
dta(
  data = sample_data(seed = 1337),
  comparator = NULL,
  benchmark = 0.5,
  alpha = 0.05,
  alternative = c("greater", "two.sided", "less"),
  adjustment = c("none", "bonferroni", "maxt", "bootstrap"),
  transformation = c("none", "logit"),
  regu = FALSE,
  pars = list(),
  ...
)
}
\arguments{
\item{data}{nxm binary matrix or data.frame (n observations of m binary decisions).
Data should have values 0 (incorrect prediction) or 1 (correct prediction).
\link{compare} provides a simple way to match predictions against true labels.
Alternatively, data can be a list of such arrays (all with m columns) defining different subsamples.}

\item{comparator}{numeric vector}

\item{benchmark}{value to compare against (RHS)}

\item{alpha}{numeric (default: 0.05), in unit interval}

\item{alternative}{character, specify alternative hypothesis}

\item{adjustment}{character, specify type of statistical adjustment taken to address multiplicity}

\item{transformation}{character, define transformation to ensure results
(e.g. point estimates, confidence limits) lie in unit interval}

\item{regu}{numeric vector of length 3, specify type of shrinkage. TODO: DETAILS
Alternatively, logical of length one (TRUE := c(2, 1, 1/2), FALSE := c(0, 0, 0))}

\item{pars}{further parameters given as named list}

\item{...}{additional named parameters}
}
\value{
DTAmcResults object, which is a list of analysis results
}
\description{
Assess classification accuracy stratified by subpopulation, in the simplest case
diseased and healthy individuals.
}
