% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data_prob}
\alias{generate_data_prob}
\title{Sample binary data (one sample)}
\usage{
generate_data_prob(n = 100, prob = c(0.8, 0.8), R = diag(length(prob)))
}
\arguments{
\item{n}{integer, sample size}

\item{prob}{numeric, vector with marginal success probabilities}

\item{R}{matrix, square correlation matrix}
}
\description{
Sample binary data (one sample)
}
