% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{sample_data}
\alias{sample_data}
\title{Sample binary data}
\usage{
sample_data(
  n = 200,
  prev = c(0.5, 0.5),
  random = FALSE,
  m = 10,
  method = c("roc", "lfc", "prob"),
  seed = NULL,
  pars = list(),
  ...
)
}
\arguments{
\item{n}{overall sample size}

\item{prev}{prevalence}

\item{random}{random sampling (TRUE) or fixed group sample sizes}

\item{m}{number of models}

\item{method}{}
}
\value{

}
\description{
Sample binary data
}
