% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data_lfc.R
\name{sample_data_lfc}
\alias{sample_data_lfc}
\title{Sample binary data (LFC model)}
\usage{
sample_data_lfc(
  n = 100,
  prev = c(0.5, 0.5),
  random = FALSE,
  m = 10,
  se = 0.8,
  sp = 0.8,
  B = round(m/2),
  L = 1,
  Rse = diag(rep(1, m)),
  Rsp = diag(rep(1, m)),
  modnames = paste0("model", 1:m),
  seed = NULL,
  ...
)
}
\arguments{
\item{n}{}

\item{prev}{}

\item{random}{}

\item{m}{}

\item{se}{}

\item{sp}{}

\item{L}{}

\item{Rse}{}

\item{Rsp}{}

\item{seed}{}

\item{...}{}

\item{b}{}
}
\value{

}
\description{
Sample binary data (LFC model)
}
