% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{sample_data_prob}
\alias{sample_data_prob}
\title{Sample binary data (one sample)}
\usage{
sample_data_prob(n = 100, prob = c(0.8, 0.8), R = diag(length(prob)))
}
\arguments{
\item{n}{}

\item{R}{}
}
\description{
Sample binary data (one sample)
}
