% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data_roc.R
\name{sample_data_roc}
\alias{sample_data_roc}
\title{Sample binary data (ROC model)}
\usage{
sample_data_roc(
  n = 100,
  prev = c(0.5, 0.5),
  random = FALSE,
  m = 10,
  auc = seq(0.85, 0.95, length.out = 5),
  rho = c(0.25, 0.25),
  e = 10,
  k = 100,
  delta = 0,
  modnames = paste0("model", 1:m),
  seed = NULL,
  corrplot = FALSE,
  ...
)
}
\arguments{
\item{n}{}

\item{prev}{}

\item{random}{}

\item{m}{}

\item{auc}{}

\item{rho}{}

\item{e}{}

\item{k}{}

\item{delta}{}

\item{modnames}{}

\item{seed}{}

\item{corrplot}{}

\item{...}{}
}
\value{

}
\description{
Sample binary data (ROC model)
}
