% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dta.R
\name{study_dta}
\alias{study_dta}
\title{Analysis for Diagnostic Test Accuracy studies with multiple index tests.}
\usage{
study_dta(
  data,
  contrast = define_contrast("raw"),
  benchmark = 0.75,
  alpha = 0.025,
  alternative = c("greater", "two.sided", "less"),
  adjustment = c("none", "bonferroni", "maxt", "bootstrap", "mbeta"),
  co_primary = TRUE,
  transformation = c("none", "logit"),
  regu = FALSE,
  pars = list(),
  ...
)
}
\arguments{
\item{data}{list of n_g x m binary matrix or data.frame (n_g observations of m binary decisions),
g is the index of subgroups/classes, usually created via \code{\link{compare}}.}

\item{contrast}{\code{DTAmc_contrast} object, specified via \code{\link{define_contrast}}}

\item{benchmark}{value to compare against (RHS), should have same length as data.}

\item{alpha}{numeric, significance level (default: 0.05)}

\item{alternative}{character, specify alternative hypothesis}

\item{adjustment}{character, specify type of statistical adjustment taken to address multiplicity}

\item{co_primary}{logical, indicates if co-primary endpoint analysis is performed (TRUE; default)
or if accuracy of all subgroups}

\item{transformation}{character, define transformation to ensure results
(e.g. point estimates, confidence limits) lie in unit interval ("none" (default) or "logit")}

\item{regu}{numeric vector of length 3, specify type of shrinkage.
Alternatively, logical of length one (TRUE := c(2, 1, 1/2), FALSE := c(0, 0, 0))}

\item{pars}{further parameters given as named list}

\item{...}{additional named parameters}
}
\value{
DTAmc_results object, which is a list of analysis results
}
\description{
Assess classification accuracy of multiple classifcation rules stratified
by subpopulation, in the simplest case diseased and healthy individuals.
}
\details{
Adjustment methods:
\itemize{
\item "none" (default): no adjustment for multiplicity
\item "bonferroni": Bonferroni adjustment
\item "maxt": maxT adjustment
\item "bootstrap": Bootstrap approach, whereby several details can be set via the
'pars' argument (see below)
\item "mbeta": Bayesian approach
}

Additional parameters to be adjusted via 'pars=list(par1=val1, par2=val2, ...)', e.g. list(type="pairs", nboot=10000), alternatively via '...'
\itemize{
\item type: "pairs" (default) or "wild" = type of bootstrap
\item nboot = number of bootstrap draws (default 5000)
\item res_tra = 0,1,2 or 3 = type of residual transformation of wild boostrap (default = 0: no transformation)
(see https://www.math.kth.se/matstat/gru/sf2930/papers/wild.bootstrap.pdf)
}
}
\examples{
#
data <- sample_data_roc()
study_dta(data)
}
