% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{threshold}
\alias{threshold}
\title{Threshold continuous (bio)markers}
\usage{
threshold(markers, cutoffs = rep(0, ncol(markers)), map = 1:ncol(markers))
}
\arguments{
\item{markers}{numeric matrix of continuous (bio)markers to be thresholded.
Assume an (n x r) matrix with n observations (subjects) of m continuous markers.}

\item{cutoffs}{numeric matrix of dimension L x K. Each row of cutoffs defines a split
into K+1 distinct categories. Each row must contain distinct values. In the simplest case,
cutoffs is a single column matrix whereby is row defines a binary split (<=t vs. >t).
In this case (K=1), cutoffs can also be a numeric vector.}

\item{map}{integer vector of length L with values in 1:r, whereby r = ncol(markers).
map_l gives the value which column of markers should be tresholded by ...}
}
\value{
numeric (n x L) matrix with categorical outcomes after thresholding
}
\description{
This function allows to split continuous (bio)markers into two or more categories
by specifying one or more cutoff values.
}
