% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{define_contrast}
\alias{define_contrast}
\title{define a contrast (matrix) to specify exact hypothesis system}
\usage{
define_contrast(type = c("raw", "Dunnett", "Tukey"), comparator = NA)
}
\arguments{
\item{type}{c("Raw", "Dunnett", "Tukey")}

\item{comparator}{either integer (index of comparator, cannot be larger than ncol(data[\link{1}]))
or character (name of rule to use for )}
}
\value{

}
\description{
define a contrast (matrix) to specify exact hypothesis system
}
