% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{define_contrast}
\alias{define_contrast}
\title{define a contrast (matrix) to specify exact hypothesis system}
\usage{
define_contrast(type = c("Raw", "Dunnett", "Tukey"), base = 1)
}
\arguments{
\item{type}{c("Raw", "Dunnett", "Tukey")}

\item{base}{}
}
\value{

}
\description{
define a contrast (matrix) to specify exact hypothesis system
}
