% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dta.R
\name{dta}
\alias{dta}
\title{Analysis for Diagnostic Test Accuracy studies with multiple index tests.}
\usage{
dta(
  data = sample_data(seed = 1337),
  contrast = define_contrast("Raw"),
  benchmark = 0.5,
  alpha = 0.05,
  alternative = c("greater", "two.sided", "less"),
  adjustment = c("none", "bonferroni", "maxt", "bootstrap", "mbeta"),
  transformation = c("none", "logit"),
  regu = FALSE,
  pars = list(),
  ...
)
}
\arguments{
\item{data}{nxm binary matrix or data.frame (n observations of m binary decisions).
Data should have values 0 (incorrect prediction) or 1 (correct prediction).
\link{compare} provides a simple way to match predictions against true labels.
Alternatively, data can be a list of such arrays (all with m columns) defining different subsamples.}

\item{contrast}{specify contrast via \link{define_contrast}}

\item{benchmark}{value to compare against (RHS), should have same length as data.}

\item{alpha}{numeric, significance level (default: 0.05)}

\item{alternative}{character, specify alternative hypothesis}

\item{adjustment}{character, specify type of statistical adjustment taken to address multiplicity}

\item{transformation}{character, define transformation to ensure results
(e.g. point estimates, confidence limits) lie in unit interval ("none" (default) or "logit")}

\item{regu}{numeric vector of length 3, specify type of shrinkage.
Alternatively, logical of length one (TRUE := c(2, 1, 1/2), FALSE := c(0, 0, 0))}

\item{pars}{further parameters given as named list}

\item{...}{additional named parameters}
}
\value{
DTAmcResults object, which is a list of analysis results
}
\description{
Assess classification accuracy of multiple classifcation rules stratified
by subpopulation, in the simplest case diseased and healthy individuals.
}
\details{
Adjustment methods:
\itemize{
\item "none" (default): no adjustment for multiplicity
\item "bonferroni": Bonferroni adjustment
\item "maxt": maxT adjustment
' - "mbeta": Bayesian approach
\item "bootstrap": Bootstrap approach, whereby several details can be set via the
'pars' argument (see below)
}

Additional parameters to be adjusted via 'pars=list(par1=val1, par2=val2, ...)', e.g. list(type="pairs", nboot=10000), alternatively via '...'
\itemize{
\item type: "pairs" (default) or "wild" = type of bootstrap
\item nboot = number of bootstrap draws (default 5000)
\item res_tra = 0,1,2 or 3 = type of residual transformation of wild boostrap (default = 0: no transformation)
(see https://www.math.kth.se/matstat/gru/sf2930/papers/wild.bootstrap.pdf)
}
}
