% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dta.R
\name{dta}
\alias{dta}
\title{dta}
\usage{
dta(
  data = sample_data(),
  test = c("univariate", "bonferroni", "maxt", "bootstrap"),
  alpha = 0.05,
  alternative = c("two.sided", "less", "greater"),
  regu = FALSE,
  pars = list(),
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
dta
}
