% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{sample_data}
\alias{sample_data}
\title{Sample binary data}
\usage{
sample_data(
  n = 200,
  prev = c(0.5, 0.5),
  random = FALSE,
  m = 10,
  method = c("roc", "lfc", "prob"),
  pars = list(),
  ...
)
}
\arguments{
\item{n}{integer, overall sample size}

\item{prev}{numeric, vector of class prevalences (adding up to 1)}

\item{random}{logical, random sampling (TRUE) or fixed group sample sizes}

\item{m}{integer, number of models}

\item{method}{character, either "roc", "lfc" (multiple subgroups) or "prob" (no subgroups)}

\item{pars}{list, containing further named parameters passed to \code{\link{sample_data_roc}},
\code{\link{sample_data_lfc}} or \code{\link{sample_data_prob}}}

\item{...}{further named parameters passed}
}
\value{
generated binary data (possibly stratified for subgroups)
}
\description{
Sample binary data
}
\examples{
sample_data()
}
