% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{sample_data}
\alias{sample_data}
\title{sample_data}
\usage{
sample_data(
  n = 200,
  m = 10,
  prev = c(0.5, 0.5),
  random = FALSE,
  method = "roc",
  pars = list(),
  ...
)
}
\arguments{
\item{n}{}

\item{m}{}

\item{prev}{}

\item{random}{}

\item{method}{}

\item{pars}{}

\item{...}{}
}
\value{

}
\description{
sample_data
}
