% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{sample_data_roc}
\alias{sample_data_roc}
\title{Title}
\usage{
sample_data_roc(
  ng = c(100, 300),
  m = 10,
  auc = seq(0.85, 0.95, length.out = 5),
  rho = c(0.25, 0.25),
  delta = 0,
  e = 10,
  k = 100,
  corrplot = FALSE,
  ...
)
}
\arguments{
\item{ng}{}

\item{m}{}

\item{auc}{}

\item{rho}{}

\item{delta}{}

\item{e}{}

\item{k}{}

\item{corrplot}{}

\item{...}{}
}
\value{

}
\description{
Title
}
